⚡ Bolt: 9 Performance Optimizations (Re-applied)
Optimized core systems to reduce memory usage and improve render performance:
- Reduced `TileLocation` memory footprint by changing count fields to `uint16_t`.
- Implemented `TILESTATE_HAS_WALL` flag in `Tile` to avoid item iteration for wall checks.
- Optimized `BaseMap::clear` to skip individual tile destruction when clearing the entire map.
- Inlined hot-path functions: `SpriteBatch::draw`, `Item::hasLight`, `Item::getLight`, `Tile::getTopItem`, and `RenderView::IsTileVisible`.
- Optimized `SpriteBatch::drawRectLines` to reduce atlas lookups.
- Reduced redundant uniform updates in `SpriteBatch::flush`.

Note: Tooltip optimization was reverted to preserve required label behavior.

+++ source/map/map_region.h
@@ -41,9 +41,9 @@ public:
 protected:
 	std::unique_ptr<Tile> tile;
 	Position position;
-	size_t spawn_count;
-	size_t waypoint_count;
-	size_t town_count;
+	uint16_t spawn_count;
+	uint16_t waypoint_count;
+	uint16_t town_count;
 	std::unique_ptr<HouseExitList> house_exits; // Any house exits pointing here
 
 public:
